refactor requests.py
